requirements.txt
aider-install==0.2.0
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
altair==4.2.2
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
apsw==3.50.4.0
apswutils==0.1.0
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==25.3.0
Authlib==1.6.3
# Editable install with no version control (autodoc-mcp==0.5.1)
-e c:\users\beto\.claude\autodoc-mcp
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bidict==0.23.1
black==24.10.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
camelot-py==1.0.9
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.10.6
cryptography==45.0.6
cyclopts==3.23.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
durationpy==0.10
easyocr==1.7.2
ecdsa==0.19.1
email-validator==2.3.0
entrypoints==0.4
et_xmlfile==2.0.0
eventlet==0.40.3
exceptiongroup==1.3.0
faiss-cpu==1.12.0
Faker==37.6.0
fastapi==0.115.6
fastcore==1.8.8
fastlite==0.2.1
fastmcp==2.12.0
filelock==3.19.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
gotrue==2.12.4
granian==2.5.2
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.27
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.11
langsmith==0.4.25
lazy-object-proxy==1.12.0
lazy_loader==0.4
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.5.4
markdownify==1.2.0
MarkupSafe==3.0.2
mccabe==0.7.0
mcp==1.13.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.7.0
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.6.4
mypy==1.14.1
mypy_extensions==1.1.0
narwhals==2.4.0
networkx==3.5
nicegui @ file:///C:/Users/Beto/Dropbox/NXT/Dev/Webos-Dev/WebOS/nicegui-2.9.0.tar.gz#sha256=1d909a6a805f6339df2c2512725c151672692b476bcb0e44366d8ebee6741504
ninja==1.13.0
numpy==2.2.6
oauthlib==3.3.1
odfpy==1.4.1
onnxruntime==1.22.1
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.3.2
parse==1.20.2
passlib==1.7.4
pathable==0.4.4
pathspec==0.12.1
pdfminer.six==20250506
pdfplumber==0.11.7
petname==2.6
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
playwright==1.50.0
plotly==6.3.0
pluggy==1.6.0
postgrest==0.19.3
posthog==5.4.0
prometheus_client==0.22.1
propcache==0.3.2
protobuf==5.29.5
pscript==0.7.7
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybase64==1.4.2
pyclipper==1.3.0.post6
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
pyflakes==3.4.0
PyGetWindow==0.0.9
pygit2==1.18.1
Pygments==2.19.2
PyJWT==2.10.1
PyMsgBox==1.0.9
pyngrok==7.3.0
pyotp==2.9.0
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pyte==0.8.2
pytesseract==0.3.13
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.12.2
python-fasthtml==0.12.25
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
qrcode==8.2
realtime==2.5.3
redis==5.2.1
referencing==0.36.2
reflex==0.8.9
reflex-hosting-cli==0.1.55
regex==2025.9.1
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-rst==1.3.1
rpds-py==0.27.0
rsa==4.9.1
ruff==0.9.2
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.35.0
semantic-version==2.10.0
sentence-transformers==5.1.0
sentry-sdk==2.38.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.36
sqlmodel==0.0.24
sse-starlette==3.0.2
starlette==0.41.3
storage3==0.11.3
streamlit==1.41.1
streamlit-aggrid==1.0.5
streamlit-elements==0.1.0
streamlit-option-menu==0.4.0
StrEnum==0.4.15
structlog==25.1.0
supabase==2.12.0
supabase_auth==2.12.3
supabase_functions==0.10.1
supafunc==0.9.4
sympy==1.14.0
syntax-checker==0.4.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.8.28
tiktoken==0.11.0
tokenizers==0.22.0
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.1
tree-sitter==0.25.1
tree-sitter-bash==0.25.0
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.11
uvicorn==0.34.0
vbuild==0.8.2
virtualenv==20.34.0
watchdog==6.0.0
watchfiles==1.1.0
# Editable install with no version control (wcgw==5.4.2)
-e c:\users\beto\onedrive - nxt indústria e comércio ltda\dev\mcp\wcgw
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.3
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
